Revert "Merge pull request #17 from William-Fraser/Upgrade-Pip-Icons-and-Layout-Edits"
This reverts commit 089a1fa12fe8536e039b08bfcce96c33272f813e, reversing
changes made to d4d5d5317e3bc4b3170ff41fcde96f0baafe5251.

+++ Assets/Scripts/Stats.cs
@@ -8,12 +8,12 @@ public class Stats : MonoBehaviour
 {
     public bool alive = true;
     public CharacterType type;
-    public float maxHealth;
+    public int maxHealth;
     public float health;
     public float speedModifier;
     public float damageModifier;
     public float attackSpeedModifier;
-    public float maxEnergy;
+    public int maxEnergy;
     public float energy;
     public int displayedEnergy; // used to display in place of energy to show better looking scale
     public float collectedMoney; // money collected in a single run
